+++ packages/web/src/app/config/screen/platform-connection.tsx
@@ -12,6 +12,7 @@ import { Container, Test, Title } from "../layout";
 import { RadioGroup } from "../../../components/RadioGroup";
 import { useWatch } from "react-hook-form";
 import Textarea from "../../../new-components/fat-fields/Textarea";
+import Subscription from "../Subscription";
 
 function PlatformConnection() {
   const dispatch = useFeatureDispatch();
@@ -22,6 +23,7 @@ function PlatformConnection() {
   } = useFeatureSelector((state) => state.config);
 
   const platformAuthType = useWatch({ name: "platformAuthType" });
+  const anondToken = useWatch({ name: "anondToken" });
 
   return (
     <>
@@ -37,9 +39,10 @@ function PlatformConnection() {
         />
 
         {platformAuthType === "anond-token" && (
-          <div>
+          <>
             <Textarea label="Freemium token" name="anondToken" />
-          </div>
+            {anondToken !== "" && <Subscription token={anondToken} />}
+          </>
         )}
 
         {platformAuthType === "api-token" && (
@@ -69,7 +72,7 @@ function PlatformConnection() {
 const schema = z.object({
   platformAuthType: z.enum(["anond-token", "api-token"]),
   platformUrl: z.string().url().startsWith("https://"),
-  anondToken: z.string(),
+  anondToken: z.string().trim(),
   platformApiToken: z
     .string()
     .regex(