+++ packages/web/src/app/config/store.ts
@@ -3,6 +3,7 @@ import {
   ListenerMiddlewareInstance,
   StateFromReducersMapObject,
 } from "@reduxjs/toolkit";
+import { setupListeners } from "@reduxjs/toolkit/query";
 import { Webapp } from "@xliic/common/webapp/config";
 import { TypedUseSelectorHook, useDispatch, useSelector } from "react-redux";
 import logger from "redux-logger";
@@ -15,10 +16,17 @@ import config, {
   showCliDownload,
 } from "../../features/config/slice";
 import theme, { changeTheme, ThemeState } from "../../features/theme/slice";
+import client from "../../features/http-client/slice";
+
+import { platformApi } from "../../features/http-client/platform-api";
+import { freemiumdApi } from "../../features/http-client/freemiumd-api";
 
 const reducer = {
   theme,
   config,
+  client,
+  [platformApi.reducerPath]: platformApi.reducer,
+  [freemiumdApi.reducerPath]: freemiumdApi.reducer,
 };
 
 export const messageHandlers: Webapp["webappHandlers"] = {
@@ -29,17 +37,26 @@ export const messageHandlers: Webapp["webappHandlers"] = {
   showScandManagerConnectionTest,
   showCliTest,
   showCliDownload,
+  showHttpError: () => null,
+  showHttpResponse: () => null,
 };
 
-export const initStore = (listenerMiddleware: ListenerMiddlewareInstance, theme: ThemeState) =>
-  configureStore({
+export const initStore = (listenerMiddleware: ListenerMiddlewareInstance, theme: ThemeState) => {
+  const store = configureStore({
     reducer,
     middleware: (getDefaultMiddleware) =>
-      getDefaultMiddleware().prepend(listenerMiddleware.middleware).concat(logger),
+      getDefaultMiddleware()
+        .prepend(listenerMiddleware.middleware)
+        .concat(logger, platformApi.middleware, freemiumdApi.middleware),
     preloadedState: {
       theme,
     },
   });
+
+  setupListeners(store.dispatch);
+
+  return store;
+};
 
 export type RootState = StateFromReducersMapObject<typeof reducer>;
 export type AppDispatch = ReturnType<typeof initStore>["dispatch"];