+++ packages/web/src/core/http-client/mock-client.ts
@@ -8,3 +8,7 @@ export type MockHttpResponseType = typeof MockHttpResponse;
 export type MockHttpClient = (
   request: HttpRequest
 ) => Promise<NullableResult<MockHttpResponseType, HttpError>>;
+
+export function mockHttpClient(): MockHttpClient {
+  return async () => [MockHttpResponse, undefined];
+}